+++ CHANGES.md
@@ -2,5 +2,6 @@
 
 
 ## Unreleased
+- Added decoders for Airrohr, Tasmota, and TTS/TTN from Kotori DAQ
 
 ## 2024/07/15 0.0.1